move query to queries.js
