+++ lib/queries.js
@@ -176,6 +176,36 @@ export const measureQuery = `
 }
 `;
 
+// samenhang queries
+export function getSamenhangData(thema, category, subCategory) {
+  const queries = {
+    beleid: `
+    *[_type == "measure" && thema == "${thema}" && ${category} == true && "beleid" in ${subCategory}][]{
+      titel,
+      ${category},
+      ${subCategory},
+      overheidslaag,
+      "slug": slug.current,
+    }
+    `,
+    strategie: `*[_type == "measure" && thema == "${thema}" && ${category} == true && "strategie" in ${subCategory}][]{
+      titel,
+      ${category},
+      ${subCategory},
+      overheidslaag,
+      "slug": slug.current,
+    }`,
+    contracten: `*[_type == "measure" && thema == "${thema}" && ${category} == true && "contracten" in ${subCategory}][]{
+      titel,
+      ${category},
+      ${subCategory},
+      overheidslaag,
+      "slug": slug.current
+    }`,
+  };
+  return queries;
+}
+
 // Query for welke overheid pages
 export function creatQuery(thema) {
   const welkQueries = {
